requirements.txt
Flask==3.1.3
docker==7.1.0
PyYAML==6.0.3
